Split EventsHeader components into their own files - Remove some comments
