+++ src/events/components/EventsHeader/index.tsx
@@ -1,12 +1,9 @@
-import classNames from 'classnames';
 import ToggleSwitch from 'common/components/ToggleSwitch';
 import React from 'react';
 import { EventView } from '../../models/Event';
+import ChangeViewIconButton from './ChangeViewIconButton';
 
-import CalendarViewIcon from './CalendarViewIcon';
 import style from './eventsHeader.less';
-import ImageViewIcon from './ImageViewIcon';
-import ListViewIcon from './ListViewIcon';
 
 export interface IProps {
   toggleAccessible: () => void;
@@ -15,45 +12,6 @@ export interface IProps {
   view: EventView;
   availableViews: number[];
 }
-
-export interface IPropsChangeViewIconButton {
-  changeView: (view: EventView) => void;
-  viewType: number;
-  view: EventView;
-}
-
-export interface IPropsViewIcon {
-  viewType: EventView;
-}
-
-const ViewIcon = ({ viewType }: IPropsViewIcon) => {
-  switch (viewType) {
-    case EventView.CALENDAR: {
-      return <CalendarViewIcon />;
-    }
-    case EventView.LIST: {
-      return <ListViewIcon />;
-    }
-    case EventView.IMAGE: {
-      return <ImageViewIcon />;
-    }
-    default: {
-      return <div />;
-    }
-  }
-};
-
-const ChangeViewIconButton = ({ viewType, changeView, view }: IPropsChangeViewIconButton) => (
-  <div
-    className={classNames(style.choice, {
-      [style.choiceActive]: view === viewType,
-    })}
-    onClick={() => changeView(viewType)}
-    tabIndex={0}
-  >
-    <ViewIcon viewType={viewType} />
-  </div>
-);
 
 const EventsHeader = ({ toggleAccessible, changeView, accessible, view, availableViews }: IProps) => (
   <div className={style.grid}>
